+++ docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ### Added
 
-- Improved error handling for bulk actions. Now they are reporting the index and entity that is failing.
+- Improved error handling for bulk actions in Python 3.11+. Now they are reporting the index and entity that is failing.
 
 ## v2.11.0 - 2026-08-20